Finished getting the base case signature check working for SqrlValidate
This will now check that a nonce exists in the database (if configured), check that the signed URL security matches the expected (qrl vs. sqrl), check that the IPs match if the enforce option is set, and check that the signature matches the URL requested.

Still to implement are tests to check how stale the nut is and also to store and retrieve user keys from the database.