Added stale nonce handling
A nonce can now be checked for age. A new configuration setting was added to the sample JSON file allowing a user to set the number of minutes before a nonce is marked as stale. Added a check and test to the SqrlValidate class to check this age against the created value stored in the database.