Updated the request handling to match the current SQRL specs
The new spec requires a full two round trips from clients to the server to perform any action. This required a near complete re-write of the RequestHandler object as well as updates to the validator. Current unit tests will not necessarily function however the new SqrlClientInteractionsIntegrationTest will lay out the primary interactions and unit tests will be built after we have the interactions working as expected.