Added handling and tests for automated account creation
When a user attempts to log in to a site that allows anonymous accounts with a public key that they have not used on that site. The site will now prompt the user to create a new account and store the new public key on user confirmation.

Also added a signature validator that always returns true for testing purposes.