Major refactoring
Brought the request handler up to date with the current version of the spec. 

As a part of this, and in part based on some discussions in the GRC newsgroups, I have also refactored some of the storage and valdiation processing. 

All validation: nut, server value, and signatures, will now be processed in the SqrlValidateInterface. The request handler will primarily only care about pass/fail states, and respond to the client appropriately.

Additionally, in order to allow more flexibility in the storage, including framework specific handling and to allow a choice between stateful and semi-stateless nut handling, SqrlStore will be removed (but recreated as a potential example) and the core library will only ship with a core SqrlStoreInterface that any framework or user will need to implement, or an abstract StatelessSqrlStore that will handle some of the intricacies of creating semi-stateless nuts.

Because of these changes, I have been able to completely remove the slow integration tests that were testing the client/server responses, and now am only doing unit tests on the SqrlRequestHandler (and soon the Generate and Validate objects)