testing fixes
updated handling for the initial identity creation to not require a URS

Updated the validate handling when checking the TIF to handle properly when a storage entity returns a string instead of an integer.

updated the sample config to include a leading / in the authentication path.

Dropped the sample SQL file that no longer matches the example server. The stateful storage example includes a sample set of SQL scripts.