Dusting off this project
Upped the minimum PHP version to PHP 7.2 since it includes Sodium in core (eventually I plan to break some of this off to support older versions with the Sodium PECL library)

updated PHPUnit to the newest version and fixed the tests to work with it.

Other minor code clean up.

still to do: static type hint everything. other PHP 7 improvements. Check the up to date spec